Python/FlaskApp/requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.23.1
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
